[transform] Add new demo-transform package for mixins, runners, and pipelines (not compiler-specific)
