+++ .env
@@ -1,0 +1,10 @@
+DEVELOPMENT.DB_URL='http://localhost:7000'
+DEVELOPMENT.LOG_LEVELS='info,warn,error'
+DEVELOPMENT.LOG_OUT='console'
+DEVELOPMENT.DEPLOYER_ADDRESS='0x54c72f1a56863ac3c14592c547b72961ee6f42b6'
+DEVELOPMENT.DEPLOYER_PASSWORD='fb0c1fd47190041bab428069eb02d0c7af71c7c084d0eafd63d1714acc278142'
+
+TEST.LOG_LEVELS='info,warn,error'
+TEST.LOG_OUT='console'
+TEST.DEPLOYER_ADDRESS="0xb3bef685fced7a2fb727b4ee51db847ef7b996f2"
+TEST.DEPLOYER_PASSWORD="6fa804606f7a5d0036dbd1cb821ea32716210758636f1a6950c827b851bf92d4"
