+++ packages/depart/index.js
@@ -1,7 +1,5 @@
 const departure = require('./src/departure')
-const runner = require('./src/runner')
 
 module.exports = {
   ...departure,
-  ...runner,
 }
