+++ packages/depart/tests/depart.spec.js
@@ -12,7 +12,8 @@ const { isContract, isCompile, isLink, isDeploy } = require('demo-contract')
 const { getImmutableKey, setImmutableKey } = require('demo-utils')
 
 const { wallet } = require('demo-keys')
-const { run, argListMixin, deployerMixin, departMixin } = require('..')
+const { run, argListMixin, deployerMixin } = require('demo-transform')
+const { departMixin } = require('..')
 
 describe( 'Departures', () => {
   
@@ -63,7 +64,7 @@ describe( 'Departures', () => {
       return new Map({ 'result': true })
     }
 
-    finalState = (await run( departFunc, [ m0, m1, m2 ] )).toJS()
+    finalState = (await run( [ m0, m1, m2, departFunc ] )).toJS()
     LOGGER.debug('finalState', finalState) 
     assert(Map.isMap(finalState.getCompiles()))
     assert(Map.isMap(finalState.getLinks()))
