+++ packages/transform/index.js
@@ -1,0 +1,5 @@
+const runners = require('./src/runner')
+
+module.exports = {
+  ...runners,
+}
