+++ packages/transform/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "demo-transform",
+  "version": "0.3.4-alpha.2",
+  "description": "Transform pipeline for a distributed country on Ethereum",
+  "directories": {
+    "tests": "tests"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "dependencies": {
+    "demo-utils": "0.3.4-alpha.1"
+  },
+  "devDependencies": {
+    "demo-tools": "0.3.4-alpha.1"
+  },
+  "scripts": {
+    "test": "NODE_ENV=DEVELOPMENT mocha ./tests --trace-warnings --exit --colors --recursive --reporter spec -t 15000",
+    "lint": "exit 0",
+    "lint:fix": "eslint --fix -c ../../.eslintrc.json *.js tests/*.js src/*.js",
+    "cov": ". ../../.covenv; ../../node_modules/nyc/bin/nyc.js --reporter=lcov --reporter=text-summary yarn test && ../../node_modules/nyc/bin/nyc.js reporter --reporter=text-lcov | npx coveralls"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/invisible-college/democracy.git"
+  },
+  "keywords": [
+    "ethereum",
+    "javascript",
+    "framework"
+  ],
+  "author": "Paul Pham",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/invisible-college/democracy/issues"
+  },
+  "homepage": "https://github.com/invisible-college/democracy#readme",
+  "gitHead": "e6eb5b4ddb1a324c68d19be66a52b2c3f77adb6c"
+}
