+++ packages/webify/exports/api.js
@@ -76,6 +76,7 @@ api.bufferGlobal  = require('bufferGlobal')
 api.contract      = require('demo-contract')
 api.keys          = require('demo-keys')
 api.tx            = require('demo-tx')
+api.transform     = require('demo-transform')
 api.immutable     = require('immutable')
 api.chai          = require('chai')
 api.secp256k1     = require('@aztec/secp256k1')
